+++ lib/flutter_remote_logger.dart
@@ -6,3 +6,4 @@ export 'src/uploader/log_uploader.dart';
 // Also exporting default implementations if users want to extend them
 export 'src/storage/file_log_storage.dart';
 export 'src/uploader/firebase_uploader.dart';
+export 'src/uploader/supabase_uploader.dart';